+++ examples/eth-dm/src/messaging/wire.ts
@@ -55,21 +55,43 @@ export class PublicKeyMessage {
   }
 }
 
-/**
- * Direct Encrypted Message used for private communication over the Waku network.
- */
-export interface DirectMessage {
-  toAddress: string;
+export interface DirectMessagePayload {
+  toAddress: Uint8Array;
   message: string;
 }
 
-export function encode<T>(msg: T): Buffer {
-  const jsonStr = JSON.stringify(msg);
-  return Buffer.from(jsonStr, 'utf-8');
-}
+/**
+ * Direct Encrypted Message used for private communication over the Waku network.
+ */
+export class DirectMessage {
+  private static Type = new Type('DirectMessage')
+    .add(new Field('toAddress', 1, 'bytes'))
+    .add(new Field('message', 2, 'string'));
+  private static Root = new Root().define('messages').add(DirectMessage.Type);
 
-export function decode<T>(bytes: Uint8Array): T {
-  const buf = Buffer.from(bytes);
-  const str = buf.toString('utf-8');
-  return JSON.parse(str);
+  constructor(public payload: DirectMessagePayload) {}
+
+  public encode(): Uint8Array {
+    const message = DirectMessage.Type.create(this.payload);
+    return DirectMessage.Type.encode(message).finish();
+  }
+
+  public static decode(bytes: Uint8Array | Buffer): DirectMessage | undefined {
+    const payload = DirectMessage.Type.decode(
+      bytes
+    ) as unknown as DirectMessagePayload;
+    if (!payload.toAddress || !payload.message) {
+      console.log('Field missing on decoded Direct Message', payload);
+      return;
+    }
+    return new DirectMessage(payload);
+  }
+
+  get toAddress(): Uint8Array {
+    return this.payload.toAddress;
+  }
+
+  get message(): string {
+    return this.payload.message;
+  }
 }